requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
asyncpg==0.25.0
pydantic==1.8.2
pyjwt==2.3.0
logstash-async==2.4.0
aiokafka==0.7.2
redis==4.0.2
azure-identity==1.7.0
azure-keyvault-secrets==4.3.0
cryptography==36.0.0
prometheus-client==0.11.0
aiohttp==3.8.1
python-dotenv==0.19.0